SatResample

SatResample.SatResample

SatResample.Grid

SatResample.EASE2N

SatResample.EASE2S

SatResample.NSDICN

SatResample.NSDICS

SatResample.gen_ps_grid

SatResample.gen_ps_grid

SatResample.genxy

SatResample.llresample

SatResample.resample

SatResample.resample

SatResample.xyfromlonlat

SatResample.SatResample — Module

SatResample – conventiently resample satellite data to common local grids.

SatResample.Grid — Type

Grid(proj, xlim, ylim, resolution)

Grid instance holding projection, border information and resolution

SatResample.EASE2N — Method

EASE2N(n)

generate the National Snow and Ice Data Center North EASE2 grid for a given resolution n in m

SatResample.EASE2S — Method

EASE2N(n)

generate the National Snow and Ice Data Center South EASE2 grid for a given resolution n in m

SatResample.NSDICN — Method

NSIDCN(n)

generate the National Snow and Ice Data Center North polar stereographic grid for a given resolution n in m

SatResample.NSDICS — Method

NSDICS(n)

generate the National Snow and Ice Data Center South polar stereographic grid for a given resolution n in m

SatResample.gen_ps_grid — Method

gen_ps_grid(clon, clat, width, height, resolution)

generates polar stereographic grid with center coordinates of clat, clon and width and height and resolution in m. This is a special case with the North pole always rotated upwards. For arbitrary rotated grid see gen_ps_grid

SatResample.gen_ps_grid — Method

gen_ps_grid(clon, clat, width, height, resolution, rotation)

generates polar stereographic grid with center coordinates of clat, clon and width and height and resolution in m rotated by rotation degrees`.

SatResample.genxy — Method

genxy(grid)

generates grid coordinates x and y for a given grid instance

SatResample.llresample — Method

resample(lon, lat, data, grid, maxdist, nneighbor)

resample data which is lat, lon coordinates into the grid. maxdist is the maximum distance in grid units to take into account nneighbor is the number of neighbors to take into account for the resampling

SatResample.resample — Method

resample(mx, my, data, xgrid, ygrid, maxdist, nneighbors)

returns unstructured ungridded data on mx,my coordinates onto a regular 2D gridded data on xgrid, ygrid

SatResample.resample — Method

resample(mx, my, data, grid, maxdist, nneighbor)

resample data which is already in grid coordinates mx and my. maxdist is the maximum distance in grid units to take into account nneighbor is the number of neighbors to take into account for the resampling

SatResample.xyfromlonlat — Method

xyfromlonlat(lon, lat, grid)

convert lat, lon into x and y coordinates of the grid